Relocate scaffold policy + snapshot + python-csv helpers (Phase 3, batch 6)
Issue #683 (parent #680, Phase 3) batch 6: continue the
`impl Agent`-method migration to `scaffold_pipeline.rs` via the
`&Agent` / `&mut Agent` veneer pattern established in batch 5.

Migrated as free fns in `scaffold_pipeline.rs` (`pub(super)`):

* `impl ScaffoldFramework { label, scaffold_hint }` (moved with the
  enum so `scaffold_pipeline` owns its full surface)
* `empty_workspace_scaffold_policy_error(&Agent, name, args)`
* `record_scaffold_artifact_snapshot(&mut Agent, request, files)`
* `python_csv_names_from_request_and_anvil(&Agent, request)`

Promoted in `turn.rs` to `pub(super)` so the moved helpers can call
back across the module boundary:

* `extract_filename_with_suffix`
* `workspace_appears_empty` (free-fn variant)

`scaffold_pipeline.rs` imports gain `super::turn::{
extract_filename_with_suffix, workspace_appears_empty}` and
`crate::agent::prompting`.

`turn.rs` adjustments:

* The 4 method slots remain as one-line veneer shims so caller sites
  (e.g. `self.empty_workspace_scaffold_policy_error(...)`) stay
  unchanged.
* Dropped now-unused `ScaffoldArtifactSnapshot` from the file-level
  `crate::session::store::{...}` import and
  `scaffold_command_matches_framework` from the top-of-file
  `super::scaffold_pipeline::{...}` re-import; the test module
  re-imports `scaffold_command_matches_framework` via the dedicated
  `super::super::scaffold_pipeline::{...}` block.

LOC delta:

* `turn.rs`: 31,353 → 31,249 (-104 LOC)
* `scaffold_pipeline.rs`: 624 → 758 (+134 LOC)
* Cumulative `turn.rs`: 39,489 → 31,249 (-8,240 LOC, -20.9%)

`pub(super)` limited / no facade re-export (DR3-001).
`turn.rs` remains the only in-crate consumer.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
